fix: assign issues to cycle, feat: settings sidebar
