+++ apps/app/layouts/navbar/settings-sidebar.tsx
@@ -1,12 +1,13 @@
 // next
+import { ArrowLeftIcon } from "@heroicons/react/24/outline";
 import Link from "next/link";
 import { useRouter } from "next/router";
 
 type Props = {
-  links: Array<{
+  links: {
     label: string;
     href: string;
-  }>;
+  }[];
 };
 
 const SettingsSidebar: React.FC<Props> = ({ links }) => {
@@ -14,9 +15,12 @@ const SettingsSidebar: React.FC<Props> = ({ links }) => {
 
   return (
     <nav className="h-screen w-72 border-r border-gray-200">
-      <div className="h-full p-2 pt-4">
-        <h2 className="text-lg font-medium leading-5">Settings</h2>
-        <div className="mt-3">
+      <div className="h-full p-2 pl-6 mt-16">
+        <h2 className="flex items-center gap-2 text-lg font-medium leading-5">
+          <ArrowLeftIcon className="h-4 w-4" />
+          Settings
+        </h2>
+        <div className="mt-6 space-y-1">
           {links.map((link, index) => (
             <h4 key={index}>
               <Link href={link.href}>
