+++ apps/app/layouts/settings-layout.tsx
@@ -19,12 +19,11 @@ type Props = {
   children: React.ReactNode;
   noPadding?: boolean;
   bg?: "primary" | "secondary";
+  noHeader?: boolean;
   breadcrumbs?: JSX.Element;
+  left?: JSX.Element;
   right?: JSX.Element;
-  links: Array<{
-    label: string;
-    href: string;
-  }>;
+  type: "workspace" | "project";
 };
 
 const SettingsLayout: React.FC<Props> = ({
@@ -32,30 +31,80 @@ const SettingsLayout: React.FC<Props> = ({
   children,
   noPadding = false,
   bg = "primary",
+  noHeader = false,
   breadcrumbs,
+  left,
   right,
-  links,
+  type,
 }) => {
   const [isOpen, setIsOpen] = useState(false);
 
   const router = useRouter();
 
-  const { user, isUserLoading } = useUser();
+  const { activeWorkspace, activeProject, user, isUserLoading } = useUser();
 
   useEffect(() => {
     if (!isUserLoading && (!user || user === null)) router.push("/signin");
   }, [isUserLoading, user, router]);
 
+  const workspaceLinks: {
+    label: string;
+    href: string;
+  }[] = [
+    {
+      label: "General",
+      href: "#",
+    },
+    {
+      label: "Control",
+      href: "#",
+    },
+    {
+      label: "States",
+      href: "#",
+    },
+    {
+      label: "Labels",
+      href: "#",
+    },
+  ];
+
+  const sidebarLinks: {
+    label: string;
+    href: string;
+  }[] = [
+    {
+      label: "General",
+      href: `/projects/${activeProject?.id}/settings`,
+    },
+    {
+      label: "Control",
+      href: `/projects/${activeProject?.id}/settings/control`,
+    },
+    {
+      label: "Members",
+      href: `/projects/${activeProject?.id}/settings/members`,
+    },
+    {
+      label: "States",
+      href: `/projects/${activeProject?.id}/settings/states`,
+    },
+    {
+      label: "Labels",
+      href: `/projects/${activeProject?.id}/settings/labels`,
+    },
+  ];
+
   return (
     <Container meta={meta}>
       <CreateProjectModal isOpen={isOpen} setIsOpen={setIsOpen} />
       <div className="h-screen w-full flex overflow-x-hidden">
-        <Sidebar />
-        <SettingsSidebar links={links} />
+        <Sidebar collapse />
+        <SettingsSidebar links={type === "workspace" ? workspaceLinks : sidebarLinks} />
         <main className="h-screen w-full flex flex-col overflow-y-auto min-w-0">
-          <Header breadcrumbs={breadcrumbs} right={right} />
+          {noHeader ? null : <Header breadcrumbs={breadcrumbs} left={left} right={right} />}
           <div
-            className={`w-full flex-grow ${noPadding ? "" : "p-5"} ${
+            className={`w-full flex-grow ${noPadding ? "" : "p-5 px-16 pt-16"} ${
               bg === "primary" ? "bg-primary" : bg === "secondary" ? "bg-secondary" : "bg-primary"
             }`}
           >
